Add files via upload
This is a Java program that implements the classic "Snake" game. It uses the Swing graphics library to create a graphical user interface for the game.

The program defines a SnakeGame class that extends the JPanel class and implements the KeyListener and Runnable interfaces. It contains several constants and instance variables:

- WIDTH, HEIGHT, and SCALE define the size of the game window and the size of each block in the game.
- FPS defines the number of frames per second that the game runs at.
- RANDOM is an instance of the Java Random class that is used to generate random positions for the food.
- snake is a LinkedList that stores the positions of the snake's body segments.
- food is a Point object that stores the position of the food.
- direction stores the current direction that the snake is moving in.
- isRunning is a boolean flag that is used to control the game loop.

The constructor for the SnakeGame class sets up the game window by creating a JFrame object and adding the current object (this) as a key listener and a JPanel. It also initializes the snake, food, and direction, and starts a new thread to run the game loop.

The paintComponent method is called by Swing whenever the game window needs to be redrawn. It draws the black background, the red food block, and the green snake blocks using the Graphics object that is passed in as a parameter.

The keyPressed method is called by Swing whenever a key is pressed. It updates the direction variable based on the key that was pressed, as long as the new direction is not opposite to the current direction.

The run method is the main game loop. It runs continuously while the is Running flag is set to true. On each iteration of the loop, it updates the position of the snake's head based on the current direction, and checks whether the head has collided with the food or any part of the snake's body. If a collision occurs, the game updates the food position and adds a new segment to the snake's body. If a collision with the wall or the snake's body occurs, the game ends and the is Running flag is set to false. Finally, the method calls repaint to redraw the game window, and sleeps for a short period of time to limit the frame rate.